+++ src/components/solverComponents/SAT_Components/prism-SAT.js
